+++ README.md
@@ -1,9 +1,20 @@
 # crud_api
-###Install packages
+Rename please .env.example to .env
+### Clone git repository
+
+git clone https://github.com/AlexOlga/crud_api.git
+
+### Change the branch to dev
+
+git checkout dev
+
+### Install dependencies
+
 ```
 npm install
 ```
 ### Run in development mode
+
 ```
 npm run dev
 ```
