+++ index.js
@@ -1,10 +1,8 @@
 import htpp from "http";
-import { randomUUID } from "crypto";
+import { userOper } from "./src/endpointUserId.js";
+import { endpointUsers } from "./src/endpointUsers.js";
 
 const PORT = process.env.PORT || 4000;
-const users = [
-  { id: randomUUID(), username: "James Bond", age: 40, hobbies: ["cars"] },
-];
 
 const server = htpp.createServer((req, res) => {
   const { method, url } = req;
@@ -15,42 +13,10 @@ const server = htpp.createServer((req, res) => {
         "Content-type": "application/json",
       });
       return res.end(JSON.stringify({ message: "main page" }));
-      break;
+
     case "/api/users":
-      if (method === "GET") {
-        res.writeHead(200, {
-          "Content-type": "application/json",
-        });
-        return res.end(JSON.stringify(users));
-      }
-
-      if (method === "POST") {
-        let data = "";
-        req.on("data", (chunk) => {
-          data += chunk;
-        });
-        req.on("end", () => {
-          
-          if (isValidData(JSON.parse(data))) {
-            const { username, age, hobbies } = JSON.parse(data);
-            const newUser={id:randomUUID(), username, age, hobbies}
-            users.push(newUser)
-            res.writeHead(201, { "Content-Type": "application/json" });
-            return res.end(JSON.stringify({
-              message: "User create",
-            }));
-          } else {
-            res.writeHead(400, { "Content-Type": "application/json" });
-            return res.end(
-              JSON.stringify({
-                message: "Enter correct data: username — string,  age — number",
-              })
-            );
-          }
-        });
-      }
+      endpointUsers(method, req, res);
       break;
-
     default:
       if (url.startsWith("/api/users/")) {
         const id = url.split("/")[3];
@@ -65,75 +31,3 @@ const server = htpp.createServer((req, res) => {
 server.listen(PORT, () => {
   console.log("server starting");
 });
-
-function userOper(method, req, res, id) {
-  const user = validUser(id);
-  // user not found
-  if (!user) {
-    res.writeHead(404, {
-      "Content-type": "application/json",
-    });
-    return res.end(JSON.stringify({ message: "User did not found" }));
-  }
-
-  switch (method.toUpperCase().trim()) {
-    case "GET":
-      res.writeHead(200, {
-        "Content-type": "application/json",
-      });
-      return res.end(JSON.stringify(user));
-    case "PUT":
-      let data = "";
-      req.on("data", (chunk) => {
-        data += chunk;
-      });
-      req.on("end", () => {        
-        if (isValidData(JSON.parse(data))) {
-          const { username, age, hobbies } = JSON.parse(data);
-          user.age = age;
-          user.username = username;
-          user.hobbies = hobbies;
-          res.writeHead(200, { "Content-Type": "application/json" });
-          return res.end(JSON.stringify(user));
-        } else {
-          res.writeHead(400, { "Content-Type": "application/json" });
-          return res.end(
-            JSON.stringify({
-              message: "Enter correct data: username — string,  age — number",
-            })
-          );
-        }
-      });
-
-      break;
-    case "DELETE":
-       const index = users.findIndex(user=> user.id===id) ;
-      if (index !== -1) {  users.splice(index, 1);}
-      res.writeHead(204, { "Content-Type": "application/json" });
-      return res.end(JSON.stringify({ message: "user deleted" }));
-    
-    default:
-      res.writeHead(500, { "Content-Type": "application/json" });
-      return res.end(JSON.stringify({ message: "invalid operation" }));
-  }
-}
-
-function validUser(id) {
-  return users.find((el) => el.id === id);
-}
-
-function isValidData(data) {
-  if (!data.username || typeof data.username !== "string") return false;
-  if (!data.age || typeof data.age !== "number") return false;
-  if (!data.hobbies) {
-    data.hobbies = [];
-  }
-  if (typeof data.hobbies !== "object") return false;
-  let flag = true;
-  data.hobbies.forEach((el) => {
-    if (typeof el !== "string") flag = false;
-  });
-  return flag;
-}
-
-
