--- routers.js
@@ -1,6 +1,0 @@
-export class Routers {
-    constructor(){
-        this.endpoints={}
-    }
-    request (method, path, header){}
-}
