+++ src/users.js
@@ -1,0 +1,5 @@
+import { randomUUID } from "crypto";
+
+export const users = [
+    { id: randomUUID(), username: "James Bond", age: 40, hobbies: ["cars"] },
+  ];
