+++ src/utilsServer.js
@@ -1,0 +1,4 @@
+export function sendResponse(res, statusCode, body) {
+    res.writeHead(statusCode, {'Content-Type': 'application/json'});
+    res.end(JSON.stringify(body));
+}
